requirements.txt
appnope==0.1.2
arrow==1.2.2
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.1.0
blessed==1.19.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.2
django-environ==0.8.1
django-filter==21.1
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.0.0
drf-excel==1.0.0
drf-renderer-xlsx==1.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
executing==0.8.2
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
ipython==8.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
openpyxl==3.0.9
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
swagger-spec-validator==2.7.4
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.1.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.8
waitress==2.0.0
wcwidth==0.2.5
zipp==3.7.0
